move files and change naming
