+++ client/src/pages/Admin/Approved/queryGetAllApprovedProjects.graphql
@@ -1,4 +1,4 @@
-#import "../../graphql/projectFragment.graphql"
+#import "../../../graphql/projectFragment.graphql"
 
 query GetAllApprovedProjects($cursor: String = null) {
   projects: getProjects(cursor: $cursor, onlyApproved: true) {
