+++ client/src/pages/Admin/Approved/style.js
